remove more unneeded stuff

+++ src/lisp.h
@@ -4844,8 +4844,8 @@ typedef struct vterminal {
   // window height has increased) and must be deleted from the terminal buffer
   int sb_pending;
   int sb_pending_by_height_decr;
-  long linenum;
-  long linenum_added;
+  int linenum;
+  int linenum_added;
 
   int invalid_start, invalid_end; // invalid rows in libvterm screen
   bool is_invalidated;
